Important refactoring on the graphics subsystem incl. OpenGL. A lot of improvements have been made, but some models are displaying quite strangely.
git-svn-id: https://gama-platform.googlecode.com/svn/branches/GAMA_CURRENT@6504 b91ef52e-2b48-11df-9ab6-b1cdc0d4a89a

+++ msi.gama.jogl/src/msi/gama/jogl/utils/GraphicDataType/MyLine.java
@@ -1,22 +1,16 @@
 package msi.gama.jogl.utils.GraphicDataType;
 
 import java.awt.Color;
-
-import msi.gama.metamodel.shape.IShape;
-import msi.gama.util.IList;
-
-import com.vividsolutions.jts.geom.Geometry;
 import com.vividsolutions.jts.geom.LineString;
 
-
 public class MyLine {
-	
+
 	public LineString line;
-	
-	public float z;
-	
+
+	public double z;
+
 	public Color color;
-	
-	public float alpha;
+
+	public double alpha;
 
 }
